#37 added components to dashboard
